requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.8.4
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.9.4
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.27.0
atomicwrites==1.4.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
daphne==4.1.2
databases==0.6.2
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==0.59.0
django-cors-headers==4.7.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
exceptiongroup==1.2.2
fastapi==0.110.2
frozenlist==1.5.0
greenlet==3.2.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.2.0
numpy==2.0.2
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.33.2
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyserial==3.5
pytest==7.1.2
pytest-asyncio==0.18.3
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
pywin32==306
PyYAML==6.0
pyzbar==0.1.9
redis==6.1.0
requests==2.28.1
requests-oauthlib==2.0.0
rfc3986==1.5.0
service-identity==24.2.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.37.2
tomli==2.2.1
twilio==9.5.2
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.2
whitenoise==6.9.0
yarl==1.19.0
zope.interface==7.2
zxing==1.0.3
